rename create-infra-structure
